move asset url process into impl file
