+++ jobs/asseturl/interface.go
@@ -6,6 +6,6 @@ import (
 
 // JobAssetURLFetcher scrapes the asset URL from the job page if not already stored.
 type JobAssetURLFetcher interface {
-	// Process receives job records and then scrapes the asset URLs and returns it.
+	// Process receives a job record and then scrapes the asset URLs and returns it.
 	Process(job jobs.Job) (jobs.JobWithAssetURL, error)
 }
